+++ packages/orq-rc/RELEASES.md
@@ -2909,3 +2909,13 @@ Based on:
 - [python v3.13.0-rc.37] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.37] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.37 - packages/orq-rc
+
+## 2025-09-25 17:48:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.628.6 (2.716.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.38] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.38] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.38 - packages/orq-rc